==> part4-five-folds-cross-validation.ipynb <==
# Melakukan impor libraries yang diperlukan untuk pembuatan kelompok cross validation.
import pandas as pd
import time
import datetime
import random

# Mencatat waktu dimulainya keseluruhan program pembuatan kelompok cross validation.
global_start_time = time.time()

# Membaca sekaligus menampilkan isi dari data training.
df_train = pd.read_csv('../input/bdc-satriadata2021/train.csv')
df_train

# Menetapkan seed untuk reproduksibilitas.
seed = 2021

# Menetapkan jumlah kelompok cross validation.
fold_number = 5

# Memisahkan dataframe untuk dua jenis kelamin berbeda.
perempuan = df_train[df_train.iloc[:, 1] == 0].sort_values('usia').reset_index()
laki_laki = df_train[df_train.iloc[:, 1] == 1].sort_values('usia').reset_index()

# Mencatat indeks foto bermasalah.
nomor_bermasalah = [104-1, 260-1, 290-1, 295-1, 767-1]

# Menyusun daftar indeks yang tidak bermasalah.
safe_pict_list_0 = []
for element in list(perempuan['index']):
    if element not in nomor_bermasalah:
        safe_pict_list_0.append(element)

safe_pict_list_1 = []
for element in list(laki_laki['index']):
    if element not in nomor_bermasalah:
        safe_pict_list_1.append(element)

# Menyiapkan daftar pengelompokan indeks untuk keperluan cross validation.
fold_dict = {integer: [] for integer in range(5)}

# Memanggil seed untuk reproduksibilitas.
random.seed(seed)

# Membagi seluruh indeks menjadi lima sama besar.
remove_size_0 = len(safe_pict_list_0) % fold_number
iteration_size_0 = len(safe_pict_list_0) // fold_number
shuffle_list_0 = safe_pict_list_0.copy()
random.shuffle(shuffle_list_0)
removed_0 = shuffle_list_0[:remove_size_0]
for element in removed_0:
    safe_pict_list_0.remove(element)
for iteration in range(iteration_size_0):
    local_part = safe_pict_list_0[iteration*fold_number: (iteration+1)*fold_number].copy()
    random.shuffle(local_part)
    for sub_element in range(fold_number):
        fold_dict[sub_element].append(local_part[sub_element])

remove_size_1 = len(safe_pict_list_1) % fold_number
iteration_size_1 = len(safe_pict_list_1) // fold_number
shuffle_list_1 = safe_pict_list_1.copy()
random.shuffle(shuffle_list_1)
removed_1 = shuffle_list_1[:remove_size_1]
for element in removed_1:
    safe_pict_list_1.remove(element)
for iteration in range(iteration_size_1):
    local_part = safe_pict_list_1[iteration*fold_number: (iteration+1)*fold_number].copy()
    random.shuffle(local_part)
    for sub_element in range(fold_number):
        fold_dict[sub_element].append(local_part[sub_element])

# Memasukkan kembali indeks yang tersisa.
leftovers = removed_0 + removed_1
random.shuffle(leftovers)
random.shuffle(nomor_bermasalah)

for fold in range(fold_number):
    fold_dict[fold].append(leftovers[fold])
    fold_dict[fold].append(nomor_bermasalah[fold])

# Menysusun indeks gambar dan kelompok cross validationnya.
fold_index = {}
for element_0 in fold_dict:
    for element_1 in fold_dict[element_0]:
        fold_index[element_1] = element_0
sorted_fold_index = [fold_index[integer] for integer in range(len(fold_index))]

# Menempatkan indeks gambar pada tiap kelompok cross validationnya.
df_train['fold'] = sorted_fold_index

# Melipatgandakan nilai pada fold, jenis kelamin, dan usia menjadi sebanyak tiga kali lipat.
triple_fold = []
for element in sorted_fold_index:
    for iteration in range(3):
        triple_fold.append(element)

jenis_kelamin = []
for element in list(df_train['jenis kelamin']):
    for iteration in range(3):
        jenis_kelamin.append(element)

usia = []
for element in list(df_train['usia']):
    for iteration in range(3):
        usia.append(element)

# Membaca sekaligus menampilkan informasi wajah pada data training.
info_train = pd.read_csv('../input/bdc-sd2021-data-tambahan/train_facial_data.csv')
info_train['fold'] = triple_fold
info_train['jenis_kelamin'] = jenis_kelamin
info_train['usia'] = usia
info_train = info_train[info_train.iloc[:, 2] != -1].reset_index().drop('index', axis = 1)
info_train

# Mencatat informasi penting seputar target dan kelompok cross validation.
important_info = info_train[['nomor', 'urutan_gambar', 'jenis_kelamin', 'usia','fold']]
important_info.to_csv('train_target_and_fold.csv', index = False)

# Mencatat waktu berakhirnya keseluruhan program pembuatan kelompok cross validation.
global_end_time = time.time()

# Menampilkan waktu eksekusi dari keseluruhan program pembuatan kelompok cross validation.
total_execution_time = datetime.timedelta(seconds = global_end_time - global_start_time)
print("total execution time: %s" % (total_execution_time))
