==> part5-prediction-cnn-0125.ipynb <==
# Melakukan impor libraries yang diperlukan untuk membangun model dan prediksi data.
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from keras.models import Sequential
from keras.layers import Dense, Dropout, Flatten, Conv2D, MaxPool2D
from tensorflow.keras.optimizers import RMSprop
from keras.preprocessing.image import ImageDataGenerator
from keras.callbacks import ReduceLROnPlateau
import time
import datetime

# Mencatat waktu dimulainya keseluruhan program model dan prediksi data.
global_start_time = time.time()

# Menetapkan nilai seed untuk reproduksi model.
seed = 2021

np.random.seed(2021)

# Menyusun dan menampilkan data training.
X_train = pd.read_csv('../input/bdc-sd2021-train-tabular-data/train_gray.csv')
X_train = X_train.values.reshape(-1,128,128,1)

# Menyusun dan menampilkan data testing.
X_test = pd.read_csv('../input/bdc-sd2021-test-tabular-data/test_gray.csv')
X_test = X_test.values.reshape(-1,128,128,1)

# Memunculkan target prediksi.
y_train = pd.read_csv('../input/bdc-sd2021-data-tambahan/train_target_and_fold.csv')[['jenis_kelamin']]
y_train

# Melakukan pemisahan data untuk training dan validasi.
X_train, X_valid, y_train, y_valid = train_test_split(X_train, y_train, test_size = 0.1, random_state = seed)

# Membangun model.
model = Sequential()
model.add(Conv2D(filters = 32, kernel_size = (5,5),padding = 'Same', 
                 activation ='relu', input_shape = (128,128,1)))
model.add(Conv2D(filters = 32, kernel_size = (5,5),padding = 'Same', 
                 activation ='relu'))
model.add(MaxPool2D(pool_size=(2,2)))
model.add(Dropout(0.25))
model.add(Conv2D(filters = 64, kernel_size = (3,3),padding = 'Same', 
                 activation ='relu'))
model.add(Conv2D(filters = 64, kernel_size = (3,3),padding = 'Same', 
                 activation ='relu'))
model.add(MaxPool2D(pool_size=(2,2), strides=(2,2)))
model.add(Dropout(0.25))
model.add(Flatten())
model.add(Dense(256, activation = "relu"))
model.add(Dropout(0.5))
model.add(Dense(1, activation='sigmoid'))
model.compile(loss = 'binary_crossentropy', optimizer='rmsprop', metrics=['accuracy'])
model.summary()

# Mendefinisikan optimizer.
optimizer = RMSprop(learning_rate=0.001, rho=0.9, epsilon=1e-08, decay=0.0)

# Mengkompilasi model.
model.compile(optimizer = optimizer , loss = "binary_crossentropy", metrics=["accuracy"])

# Menetapkan batasan untuk learning rate.
learning_rate_reduction = ReduceLROnPlateau(monitor='val_acc', 
                                            patience=2, 
                                            verbose=1, 
                                            factor=0.5, 
                                            min_lr=0.00001)

# Menetapkan nilai epoch dan batch size.
epochs = 85
batch_size = 90

# Melakukan augmentasi data.
datagen = ImageDataGenerator(
        featurewise_center=False,  # set input mean to 0 over the dataset
        samplewise_center=False,  # set each sample mean to 0
        featurewise_std_normalization=False,  # divide inputs by std of the dataset
        samplewise_std_normalization=False,  # divide each input by its std
        zca_whitening=False,  # apply ZCA whitening
        rotation_range=10,  # randomly rotate images in the range (degrees, 0 to 180)
        zoom_range = 0.1, # Randomly zoom image
        width_shift_range=0.1,  # randomly shift images horizontally (fraction of total width)
        height_shift_range=0.1,  # randomly shift images vertically (fraction of total height)
        horizontal_flip=False,  # randomly flip images
        vertical_flip=False)  # randomly flip images

datagen.fit(X_train)

# Melakukan fitting model.
history = model.fit(datagen.flow(X_train,y_train, batch_size=batch_size),
                    epochs = epochs, validation_data = (X_valid,y_valid),
                    verbose = 2, steps_per_epoch = X_train.shape[0] // batch_size,
                    callbacks = [learning_rate_reduction])

# Memprediksi target pada data validasi dan testing.
y_valid_pred = model.predict(X_valid)
y_valid_pred = pd.DataFrame((y_valid_pred >= 0.5)*1)

y_test_pred = model.predict(X_test)
y_test_pred = pd.DataFrame((y_test_pred >= 0.5) * 1)

# Menampilkan nilai error pada data validasi.
y_valid = y_valid.reset_index().drop('index', axis = 1)
y_valid.columns = [0]
error = abs(y_valid - y_valid_pred).mean()
print('nilai error pada data validasi: ', error[0])

# Menyimpan hasil prediksi.
prediction = pd.read_csv('../input/bdc-satriadata2021/submission.csv')
prediction['jenis kelamin'] = y_test_pred
prediction.to_csv('prediction_cnn_0125.csv', index = False)
prediction

# Mencatat waktu berakhirnya keseluruhan program model dan prediksi data.
global_end_time = time.time()

# Menampilkan waktu eksekusi dari keseluruhan program model dan prediksi data.
total_execution_time = datetime.timedelta(seconds = global_end_time - global_start_time)
print("total execution time: %s" % (total_execution_time))
