==> part6b-test-facial-landmark-relative-position.ipynb <==
# Melakukan impor libraries yang diperlukan untuk ekstraksi informasi wajah pada testing images.
import numpy as np
import pandas as pd
import matplotlib.image as mpimg
import math
import cv2
import time
import datetime

# Melakukan impor MTCNN untuk deteksi wajah.
! pip install mtcnn
from mtcnn.mtcnn import MTCNN

# Mencatat waktu dimulainya keseluruhan program ekstraksi informasi wajah pada testing images.
global_start_time = time.time()

# Membaca sekaligus menampilkan id dari data testing.
df_test = pd.read_csv('../input/bdc-satriadata2021/submission.csv')
df_test

# Mencatat waktu dimulainya pembacaan testing images.
start_time = time.time()

# Mencatat index dari setiap testing images.
img_test = {}
index_test = list(df_test.iloc[:,0])
filename_master_test = '../input/bdc-satriadata2021/Testing/'

for index in index_test:
    # Membaca dan menyimpan testing images.
    img_test[index] = {}
    img_index = index + '.jpg'
    filename_test = filename_master_test + img_index
    img_test[index] = mpimg.imread(filename_test)

# Mencatat waktu berakhirnya pembacaan testing images.
end_time = time.time()

# Menampilkan waktu eksekusi dari pembacaan testing images.
execution_time = datetime.timedelta(seconds = end_time - start_time)
print("execution time: %s" % (execution_time))

# Mendeklarasikan pendeteksi wajah.
detector = MTCNN()

# Membaca sekaligus menampilkan informasi wajah pada data testing.
info_test = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_facial_data_updated.csv')
info_test

# Menentukan ukuran gambar dan margin gambar saat akan dirotasi.
adj_ratio = 0.25
pict_size = (128, 128)

# Mencatat waktu dimulainya pencatatan informasi testing images.
start_time = time.time()

# Menyiapkan ukuran penyimpanan informasi gambar.
nrows = info_test.shape[0]

# Menyiapkan dataframe untuk menampung informasi gambar.
facial_relative = pd.DataFrame()

for row in range(nrows):
    # Menyiapkan indeks gambar.
    pict_index = info_test.iloc[row, 1]
    
    # Menyiapkan gambar untuk dirotasi.
    my_image = img_test[pict_index].copy()
    x_coor, y_coor, width, height = info_test.iloc[row, 2:6]
    adj_x_coor = max(x_coor-int(adj_ratio*width), 0)
    adj_y_coor = max(y_coor-int(adj_ratio*height), 0)
    adj_width = int(width*(1+2*adj_ratio))
    adj_height = int(height*(1+2*adj_ratio))
    crop = my_image[adj_y_coor:adj_y_coor+adj_height, adj_x_coor:adj_x_coor+adj_width]
    
    # Melakukan rotasi.
    delta_x_eyes = int(info_test.iloc[row, 8] - info_test.iloc[row, 6])
    delta_y_eyes = int(info_test.iloc[row, 9] - info_test.iloc[row, 7])
    x_center = crop.shape[1] // 2
    y_center = crop.shape[0] // 2
    angle = math.degrees(math.atan(delta_y_eyes/delta_x_eyes))
    M = cv2.getRotationMatrix2D((x_center, y_center), angle, 1)
    out = cv2.warpAffine(crop, M, (crop.shape[1], crop.shape[0]))
    
    # Menentukan lokasi wajah yang baru.
    face_location = detector.detect_faces(out)
    if len(face_location):
        new_x, new_y, new_w, new_h = face_location[0]['box']
        main_face = out[new_y:new_y+new_h, new_x:new_x+new_w]
        
        # Memasukkan informasi posisi relatif fitur pada wajah.
        temp_face = []
        main_element = face_location[0]['keypoints']
        for face_element in main_element:
            temp_face_0 = (np.array(main_element[face_element]) - (new_x, new_y)) / (new_w, new_h)
            temp_face.append(list(temp_face_0))
        temp_face = list(np.array(temp_face).reshape(10))
        temp_face.append(new_h / new_w)
        
    else:
        # Definisi ulang gambar, jika deteksi wajah gagal dilakukan.
        crop = my_image[y_coor:y_coor+height, x_coor:x_coor+width]
        # Melakukan rotasi.
        delta_x_eyes = int(info_test.iloc[row, 8] - info_test.iloc[row, 6])
        delta_y_eyes = int(info_test.iloc[row, 9] - info_test.iloc[row, 7])
        x_center = crop.shape[1] // 2
        y_center = crop.shape[0] // 2
        angle = math.degrees(math.atan(delta_y_eyes/delta_x_eyes))
        M = cv2.getRotationMatrix2D((x_center, y_center), angle, 1)
        out = cv2.warpAffine(crop, M, (crop.shape[1], crop.shape[0]))
        main_face = out
        
        # Memasukkan informasi posisi relatif fitur pada wajah.
        temp_face = []
        for temp_index in range(5):
            temp_face_0 = (np.array(info_test.iloc[row, 6+2*temp_index:8+2*temp_index]) - (x_coor, y_coor)) / (width, height)
            temp_face.append(list(temp_face_0))
        temp_face = list(np.array(temp_face).reshape(10))
        temp_face.append(height / width)
    
    # Memasukkan informasi posisi relatif yang diukur ulang tanpa mengubah rasio ukuran wajah.
    adj_y = []
    for temp_index in range(len(temp_face)):
        if temp_index%2:
            adj_y.append(temp_face[temp_index] * temp_face[-1])
    temp_face += adj_y
    
    # Membuat nama kolom untuk informasi wajah.
    facial_element = ['_le', '_re', '_n', '_ml', '_mr']
    col_name = []
    for temp_element in facial_element:
        col_name.append('x' + temp_element)
        col_name.append('y' + temp_element)
    col_name.append('wh_ratio')
    for temp_element in facial_element:
        col_name.append('adj_y' + temp_element)
    
    # Mencatat koordinat relatif wajah.
    x_face = []
    y_face = []
    for temp_index in range(5):
        x_face.append(temp_face[2*temp_index])
        y_face.append(temp_face[2*temp_index+1])
    
    # Memasukkan informasi jarak antar titik penting pada wajah.
    for temp_index_0 in range(5):
        for temp_index_1 in range(temp_index_0):
            temp_name = facial_element[temp_index_1] + facial_element[temp_index_0]
            delta_x = x_face[temp_index_0] - x_face[temp_index_1]
            delta_y = y_face[temp_index_0] - y_face[temp_index_1]
            adj_delta_y = delta_y * temp_face[10]
            
            # Jarak simetris.
            sym_distance = np.sqrt(delta_x**2 + delta_y**2)
            temp_face.append(sym_distance)
            col_name.append('sym' + temp_name)
            
            # Jarak yang disesuaikan dengan rasio ukuran wajah.
            adj_distance = np.sqrt(delta_x**2 + adj_delta_y**2)
            temp_face.append(adj_distance)
            col_name.append('adj' + temp_name)
    
    # Memasukkan informasi kemiringan sudut wajah pada gambar.
    temp_face.append(abs(math.radians(angle)))
    col_name.append('abs_angle')
    
    # Mengkompilasi seluruh informasi wajah dalam suatu dataframe.
    face_df = pd.DataFrame(temp_face).T
    facial_relative = pd.concat([facial_relative, face_df], ignore_index = True)

# Menyimpan seluruh informasi gambar ke dalam format .csv.
facial_relative.columns = col_name
facial_relative.to_csv('test_facial_relative.csv', index = False)

# Mencatat waktu berakhirnya pencatatan informasi testing images.
end_time = time.time()

# Menampilkan waktu eksekusi dari pencatatan informasi testing images.
execution_time = datetime.timedelta(seconds = end_time - start_time)
print("execution time: %s" % (execution_time))

# Menampilkan informasi gambar yang sudah diperoleh.
facial_relative

# Mencatat waktu berakhirnya keseluruhan program ekstraksi informasi wajah pada testing images.
global_end_time = time.time()

# Menampilkan waktu eksekusi dari keseluruhan program ekstraksi informasi wajah pada testing images.
total_execution_time = datetime.timedelta(seconds = global_end_time - global_start_time)
print("total execution time: %s" % (total_execution_time))
