==> part8-submisi-akhir.ipynb <==
# Impor file yang diperlukan untuk submisi akhir.
import numpy as np
import pandas as pd

# Prediksi pertama.
prediction_0 = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_preds_1_03_03_0.csv').iloc[:, 0]
prediction_0

# Prediksi kedua.
prediction_1 = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_preds_1_04_01_0.csv').iloc[:, 0]
prediction_1

# Prediksi ketiga.
prediction_2 = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_preds_1_04_01_1.csv').iloc[:, 0]
prediction_2

# Prediksi keempat.
prediction_3 = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_preds_1_04_01_2.csv').iloc[:, 0]
prediction_3

# Prediksi kelima.
prediction_4 = pd.read_csv('../input/bdc-sd2021-data-tambahan/test_preds_1_04_01_4.csv').iloc[:, 0]
prediction_4

# Prediksi yang dikompilasi.
prediction = pd.DataFrame()
prediction['model_0'] = prediction_0
prediction['model_1'] = prediction_1
prediction['model_2'] = prediction_2
prediction['model_3'] = prediction_3
prediction['model_4'] = prediction_4
prediction

# Menghitung rata-rata prediksi.
final_prediction = np.round(prediction.mean(axis = 1)).astype(int)
final_prediction

# Menyimpan submisi akhir.
submission = pd.read_csv('../input/bdc-satriadata2021/submission.csv')
submission['usia'] = final_prediction
submission.to_csv('submission.csv', index = False)
submission
